Project Almost Complete
Today, I added in three tracks that created an eerie tone. Two of them were rocket ships taking off and landing and I also added in part of the Alien soundtrack. I changed the pitch in my voice so that it would sound more powerful. I used the amplify tool, the fade in tool, and the envelope tool to make the tracks flow. Overall, I think my project is nearly complete. I just need to clean it up a bit.
Here is my box folder with the files:
https://pitt.app.box.com/folder/39671293211